src/webscraping/pyproject.toml
[project]
name = "finwhiz-scraper"
version = "0.1.0"
description = "FinWhiz web scraping and data ingestion for IRS, Consumer Finance, and FINRA"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12",
    "pymupdf>=1.24",
    "requests>=2.31",
    "google-cloud-storage>=2.14",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ADD THIS SECTION:
[tool.hatch.build.targets.wheel]
packages = ["ingest_shared"] 